backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.30.1
groq>=0.9.0
networkx>=3.0
sqlparse>=0.5.0
pytest==8.2.0
pytest-asyncio==0.23.6
